Former-commit-id: f224f987c855971ff6611009711cb78d99910b7a

+++ src/game/CreatureGroups.cpp
@@ -99,7 +99,7 @@ void CreatureGroupManager::LoadCreatureFormations()
         group_member                        = new FormationInfo;
         group_member->leaderGUID            = fields[0].GetUInt32();
         uint32 memberGUID = fields[1].GetUInt32();
-        group_member->groupAI                = fields[5].GetUInt8();
+        group_member->groupAI                = fields[5].GetUInt32();
         group_member->respawn                = fields[6].GetBool();
         group_member->linkedLoot             = fields[7].GetBool();
         //If creature is group leader we may skip loading of dist/angle
