+++ src/game/InstanceSaveMgr.cpp
@@ -413,7 +413,7 @@ void InstanceSaveManager::LoadResetTimes()
         do
         {
             Field *fields = result->Fetch();
-            uint32 mapid = fields[0].GetUInt32();
+            uint32 mapid = fields[0].GetUInt16();
             if(!sObjectMgr->GetInstanceTemplate(mapid))
             {
                 TC_LOG_ERROR("misc","InstanceSaveManager::LoadResetTimes: invalid mapid %u in instance_reset!", mapid);
@@ -422,7 +422,7 @@ void InstanceSaveManager::LoadResetTimes()
             }
 
             // update the reset time if the hour in the configs changes
-            uint64 oldresettime = fields[1].GetUInt64();
+            uint64 oldresettime = fields[1].GetUInt32();
             uint64 newresettime = (oldresettime / DAY) * DAY + diff;
             if(oldresettime != newresettime)
                 CharacterDatabase.DirectPExecute("UPDATE instance_reset SET resettime = '" UI64FMTD "' WHERE mapid = '%u'", newresettime, mapid);
